initial versions of shared datastructures feature and characteristic:
feature: Single feature passed from extractor to summarizer. Contains
e.g. the dataset-id ("id"), the name of the feature ("name"), the type
("type", at the moment possible values are "enumerable" or "other") and
a value.

characteristic: Aggregates possibly multiple features. Is an object
representation of the database-entries in the
"characteristics"-collection/table.